+++ src/tools/tool-options.h
@@ -182,10 +182,6 @@ struct ToolOptions : public Options {
   void applyFeatures(Module& module) const {
     module.features.enable(enabledFeatures);
     module.features.disable(disabledFeatures);
-    // Non-default type systems only make sense with GC enabled.
-    if (!module.features.hasGC() && getTypeSystem() == TypeSystem::Nominal) {
-      Fatal() << "Nominal typing is only allowed when GC is enabled";
-    }
   }
 
 private: